singup fix

+++ controller/functions.php
@@ -49,7 +49,7 @@ function createTableLid(){
     achternaam VARCHAR(255) NOT NULL,
     email VARCHAR(255) NOT NULL UNIQUE,
     gb_datum DATE NOT NULL,
-    leeftijd INT(50) NOT NULL,
+    leeftijd SMALLINT(100) NOT NULL,
     soort_lid VARCHAR(50) NOT NULL,
     PRIMARY KEY (id_lid)
 )";
